pyproject.toml
[tool.poetry]
name = "qms"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
requests = "^2.27.1"
iexfinance = "^0.5.0"
numpy = "^1.22.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
